requirements.txt
google-generativeai==0.5.2
gtts==2.5.1
SpeechRecognition==3.10.1
pyaudio==0.2.14
python-dotenv==1.0.1
